update dependencies part 6
Updated the following dependencies:
"mocha": "10.2.0"
"should": "13.2.3"
"supertest": "6.3.3"

Updated the last of the dependencies.
The search suggest wordform feature slowed down for some reason so it was optimised.
See comment in routes\wordforms.js /search_suggest for details.
Also added more unit tests for when a lexeme or wordform starts with a possible diacritic or not.